+++ com.dotsui.core/DotsUI.Core/DotsUISystemGroups.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using DotsUI.Core.Utils;
 using Unity.Entities;
 using Unity.Transforms;
 
@@ -16,7 +17,7 @@ namespace DotsUI.Core
 
     [UpdateInGroup(typeof(SimulationSystemGroup))]
     [UpdateAfter(typeof(InputSystemGroup))]
-    public class InputHandleBarrier : EntityCommandBufferSystem
+    public class InputHandleBarrier : MicroCommandBufferSystem
     {
 
     }