+++ com.dotsui.core/DotsUI.Input/UpdateKeyboardEvents.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 using Unity.Collections;
 using Unity.Entities;
 using Unity.Jobs;
@@ -6,30 +7,41 @@ using UnityEngine;
 
 namespace DotsUI.Input
 {
-    //internal struct UpdateKeyboardEvents : IJob
-    //{
-    //    [ReadOnly] [DeallocateOnJobCompletion] public NativeArray<(NativeInputEventType, KeyboardInputBuffer)> KeyboardFrameData;
-    //    [ReadOnly] public ComponentDataFromEntity<InputSystemState> StateFromEntity;
-    //    [ReadOnly] public Entity StateEntity;
-    //    public EntityCommandBuffer Manager;
-    //    public EntityArchetype EventArchetype;
+    internal struct UpdateKeyboardEvents : IJob
+    {
+        [ReadOnly] [DeallocateOnJobCompletion] public NativeArray<NativeKeyboardInputEvent> KeyboardEvents;
+        [ReadOnly] public ComponentDataFromEntity<InputSystemState> StateFromEntity;
+        [ReadOnly] public Entity StateEntity;
+        public EntityCommandBuffer Manager;
+        public EntityArchetype EventArchetype;
 
-    //    public void Execute()
-    //    {
-    //        var state = StateFromEntity[StateEntity];
+        public void Execute()
+        {
+            var state = StateFromEntity[StateEntity];
 
-    //        if (state.SelectedEntity != default)
-    //        {
-    //            var eventEntity = Manager.CreateEntity(EventArchetype);
-    //            Manager.SetComponent(eventEntity, new KeyboardEvent()
-    //            {
-    //                Target = state.SelectedEntity
-    //            });
-    //            var buffer = Manager.SetBuffer<KeyboardInputBuffer>(eventEntity);
-    //            for (int i = 0; i < KeyboardFrameData.Length; i++)
-    //                if (KeyboardFrameData[i].Item1 == NativeInputEventType.KeyDown)
-    //                    buffer.Add(KeyboardFrameData[i].Item2);
-    //        }
-    //    }
-    //}
+            if (KeyboardEvents.Length > 0)
+            {
+                var selected = state.SelectedEntity;
+                if (selected != default)
+                {
+                    var eventEntity = Manager.CreateEntity(EventArchetype);
+                    var buff = Manager.SetBuffer<KeyboardInputBuffer>(eventEntity);
+                    Manager.SetComponent(eventEntity, new KeyboardEvent()
+                    {
+                        Target = selected
+                    });
+                    for (int i = 0; i < KeyboardEvents.Length; i++)
+                    {
+                        if (KeyboardEvents[i].EventType == NativeInputEventType.KeyDown)
+                            buff.Add(new KeyboardInputBuffer()
+                            {
+                                Character = KeyboardEvents[i].Character,
+                                EventType = KeyboardEvents[i].KbdEvent,
+                                KeyCode = KeyboardEvents[i].KeyCode
+                            });
+                    }
+                }
+            }
+        }
+    }
 }